+++ projects/framework-nuke/release_notes.md
@@ -1,9 +1,16 @@
 # ftrack Framework Nuke integration release Notes
 
+## Upcoming
+
+* [changed] When selecting the node to render for reviewable, show all nodes with the selected one on top.
+* [new] Writeable node validator, to check if a node is writeable for a reviewable to be created.
+
+
 ## 24.3.0rc2
 2024-03-26
 
 * [changed] Upgrade ftrack-framework-core, ftrack-utils and ftrack-qt version.
+
 
 ## v24.3.0rc1
 2024-03-25